+++ src/Application/Commands/CrawlCommand.php
@@ -11,6 +11,7 @@ use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Question\ConfirmationQuestion;
 
 class CrawlCommand extends Command
 {
@@ -19,6 +20,7 @@ class CrawlCommand extends Command
         $this->setName('crawl');
         $this->setDescription('Runs the crawler with specified configuration');
         $this->addArgument('config', InputArgument::REQUIRED, 'Path to the configuration file');
+        $this->addOption('force', 'f', InputOption::VALUE_OPTIONAL, 'Do not ask on potential directory override', false);
         $this->addOption('output', 'o', InputOption::VALUE_OPTIONAL, 'Path to the output directory, created next to config file by default');
     }
 
@@ -33,6 +35,15 @@ class CrawlCommand extends Command
 
         if (is_file($configurationPath) === false) {
             throw new \Exception("Could not find configuration at '{$configurationPath}'");
+        }
+
+        if (is_dir($outputPath) && $input->getOption('force') === false) {
+            $helper = $this->getHelper('question');
+            $confirmation = new ConfirmationQuestion('Directory already exists, do you want to remove it (will most probably override old results too)? [Y/n] ');
+
+            if ($helper->ask($input, $output, $confirmation) === false) {
+                return 0;
+            }
         }
 
         /** @noinspection PhpIncludeInspection */