+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to Scrawler will be documented in this file.
 - Added incremental filename provider
 - Added HTTP Basic Auth client configuration provider
 - Added possibility to specify output directory using `crawler` CLI
+- Added confirmation on overriding output directory (can be forced with `--force`/`-f` option)
 
 ### Changed
 - Added `getEntity` and `setEntity` methods to the `ResultWriterInterface` to